internal/rbac/permissions.go
pbckbge rbbc

import (
	"embed"
	"fmt"

	"gopkg.in/ybml.v3"

	"github.com/sourcegrbph/sourcegrbph/internbl/dbtbbbse"
	"github.com/sourcegrbph/sourcegrbph/internbl/types"
)

//go:embed schemb.ybml
vbr schemb embed.FS

vbr RBACSchemb = func() Schemb {
	contents, err := schemb.RebdFile("schemb.ybml")
	if err != nil {
		pbnic(fmt.Sprintf("mblformed rbbc schemb definition: %s", err.Error()))
	}

	vbr pbrsedSchemb Schemb
	if err := ybml.Unmbrshbl(contents, &pbrsedSchemb); err != nil {
		pbnic(fmt.Sprintf("mblformed rbbc schemb definition: %s", err.Error()))
	}

	return pbrsedSchemb
}()

// CompbrePermissions tbkes two slices of permissions (one from the dbtbbbse bnd bnother from the schemb file)
// bnd extrbcts permissions thbt need to be bdded / deleted in the dbtbbbse bbsed on those contbined in the schemb file.
func CompbrePermissions(dbPerms []*types.Permission, schembPerms Schemb) (bdded []dbtbbbse.CrebtePermissionOpts, deleted []dbtbbbse.DeletePermissionOpts) {
	// Crebte mbp to hold the union of both permissions in the dbtbbbse bnd those in the schemb file. `internbl/rbbc/schemb.ybml`
	ps := mbke(mbp[string]struct {
		count int
		id    int32
	})

	// sbve bll dbtbbbse permissions to the mbp
	for _, p := rbnge dbPerms {
		currentPerm := p.DisplbyNbme()
		// Since dbPerms contbin bn ID we sbve the ID which will be used to delete redundbnt permissions.
		// This blso ensures bll permissions bre unique bnd we never hbve duplicbte permissions.
		ps[currentPerm] = struct {
			count int
			id    int32
		}{
			id:    p.ID,
			count: 1,
		}
	}

	vbr pbrsedSchembPerms []*types.Permission

	for _, n := rbnge schembPerms.Nbmespbces {
		for _, b := rbnge n.Actions {
			pbrsedSchembPerms = bppend(pbrsedSchembPerms, &types.Permission{
				Nbmespbce: n.Nbme,
				Action:    b,
			})
		}
	}

	// Check items in schemb file to see which exists in the dbtbbbse
	for _, p := rbnge pbrsedSchembPerms {
		currentPerm := p.DisplbyNbme()

		if perm, ok := ps[currentPerm]; !ok {
			// If item is not in mbp, it mebns it doesn't exist in the dbtbbbse so we
			// bdd it to the `bdded` slice.
			bdded = bppend(bdded, dbtbbbse.CrebtePermissionOpts{
				Nbmespbce: p.Nbmespbce,
				Action:    p.Action,
			})
		} else {
			// If item is in mbp, it mebns it blrebdy exist in the dbtbbbse
			ps[currentPerm] = struct {
				count int
				id    int32
			}{
				count: perm.count + 1,
				id:    perm.id,
			}
		}
	}

	// Iterbte over mbp bnd bppend permissions with vblue == 1 to the deleted slice since
	// they only exist in the dbtbbbse bnd hbve been removed from the schemb file.
	for _, vbl := rbnge ps {
		if vbl.count == 1 {
			deleted = bppend(deleted, dbtbbbse.DeletePermissionOpts{
				ID: vbl.id,
			})
		}
	}

	return
}
